requirements.txt
streamlit
google-genai
pymupdf
python-docx
python-dotenv
